final/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiview-3d-reconstruction"
version = "0.1.0"
description = "A starter project for 3D object reconstruction from multiple views."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pillow>=10.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "tqdm>=4.66",
  "gradio>=5.0",
  "open3d>=0.19.0",
]

[project.scripts]
reconstruct3d-train = "reconstruction3d.train:main"
reconstruct3d-infer = "reconstruction3d.infer:main"
reconstruct3d-app = "reconstruction3d.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
